Modularize hover provider
