+++ src/providers/hover/index.js
@@ -1,0 +1,5 @@
+const { status } = require("./status");
+
+const hoverProvider = (params, documents) => status(params, documents);
+
+module.exports = { hoverProvider };
